requirements.txt
Package            Version
------------------ ---------
asgiref            3.6.0
certifi            2022.12.7
cffi               1.15.1
charset-normalizer 3.1.0
cryptography       40.0.2
defusedxml         0.7.1
Django             4.2
django-allauth     0.54.0
django-filter      23.1
idna               3.4
oauthlib           3.2.2
pip                22.3.1
pycparser          2.21
PyJWT              2.6.0
python3-openid     3.2.0
requests           2.28.2
requests-oauthlib  1.3.1
setuptools         65.6.3
sqlparse           0.4.4
urllib3            1.26.15
